requirements.txt
python-dotenv==1.0.1
requests==2.28.2
tabulate==0.9.0
urllib3==1.26.6
